# test: slightly faster user tests