This is an attempt to make the user stats tests slightly faster. Turns out that the setup function was migrated from forum v1, so it's probably better that we don't change the code too much until we've reached stable status.

With this changes, we observe ~20% improvement in the top 4 user tests:

0.82s setup tests/test_views/test_users.py::test_returns_users_stats_with_default_activity_sort 0.81s setup tests/test_views/test_users.py::test_update_user_stats 0.81s setup tests/test_views/test_users.py::test_returns_users_stats_filtered_by_user_with_default_activity_sort 0.80s setup tests/test_views/test_users.py::test_returns_users_stats_with_flagged_sort 0.76s call  tests/test_views/test_search.py::test_pagination

These slow tests were migrated from forum v1, so it's probably better that we don't change the code too much until we've reached stable status.

This was meant to address #94, but we didn't really manage to resolve the issue fully.