@regisb 2024-10-03
This is really unnecessary. We can always investigate slow tests once we are done with the mysql migration.